chore(bench): split specifier microbenches into a separate bench binary
`specifier/realistic[rw/hash-only]` (~340 instructions) and
`specifier/branches[fragment/short]` (~170 instructions) are short
enough that the cold instruction-cache fill on each CodSpeed
measurement iteration dominates the result: ~200 cycles of fixed
overhead translates to +5% to +10% deltas under any unrelated
binary-layout shift, even when the parser itself is unchanged or
faster. CodSpeed surfaces these as false-positive regressions.

Move the four `specifier/*` bench groups into their own
`benches/specifier.rs` and register it as a second `[[bench]]` in
Cargo.toml. Each `[[bench]]` runs in its own process, so the
specifier binary gets a fresh, much smaller instruction-cache
footprint instead of competing with the large `bench_resolver` code
for cache lines. The per-case Ir is unchanged — what changes is the
working-set the kernel and the L1/LL caches see before measurement
starts, which makes cold-start misses predictable across runs.

- `benches/specifier.rs`: new file. Allocator wrapper mirrors the one
  in `bench_resolver` so allocation costs are measured identically.
- `benches/resolver.rs`: drops `specifier/*` groups, helpers,
  unused imports.
- `Cargo.toml`: adds `[[bench]] name = "specifier"`.

No code paths in `Specifier::parse` are touched; this is purely
test-infrastructure stabilization.
